+++ fdb-record-layer-lucene/src/test/java/com/apple/foundationdb/record/lucene/LuceneAnalyzerTest.java
@@ -21,6 +21,7 @@
 package com.apple.foundationdb.record.lucene;
 
 import com.apple.foundationdb.record.lucene.ngram.NgramAnalyzer;
+import com.apple.foundationdb.record.lucene.synonym.SynonymAnalyzer;
 import com.google.common.collect.ImmutableSet;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.analysis.CharArraySet;
@@ -66,6 +67,15 @@ public class LuceneAnalyzerTest {
 
         tokenizeWithAnalyzer(result, input, new NgramAnalyzer(stopWords, 3, 10, false));
         Assertions.assertEquals(ImmutableSet.of("rl"), result);
+    }
+
+    @Test
+    void testSynonymAnalyzer() throws Exception {
+        String input = "Hello RL";
+        Collection<String> result = new HashSet<>();
+
+        tokenizeWithAnalyzer(result, input, new SynonymAnalyzer(null));
+        Assertions.assertEquals(ImmutableSet.of("hi", "how", "hullo", "howdy", "rl", "hello", "do", "you"), result);
     }
 
     @Test